Add SGD2MAPI extension for glide3x D2DX detection
